Fix warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
visp/tutorial/grabber/tutorial-grabber-opencv.cpp:177:47: warning: implicit conversion changes signedness: 'unsigned int' to 'int' [-Wsign-conversion]
  177 |       vpDisplay::displayText(I, I.getHeight() - 20, 10, ss.str(), vpColor::red);
      |       ~~~~~~~~~                 ~~~~~~~~~~~~~~^~~~

+++ tutorial/grabber/tutorial-grabber-opencv.cpp
@@ -174,7 +174,7 @@ int main(int argc, const char *argv[])
 
       std::stringstream ss;
       ss << "Acquisition time: " << std::setprecision(3) << vpTime::measureTimeMs() - t << " ms";
-      vpDisplay::displayText(I, I.getHeight() - 20, 10, ss.str(), vpColor::red);
+      vpDisplay::displayText(I, static_cast<int>(I.getHeight()) - 20, 10, ss.str(), vpColor::red);
       vpDisplay::flush(I);
     }
     image_queue.cancel();
